+++ tests-e2e/reference_output/test_generic_args1/test_generic_args1.d.ts
@@ -24,31 +24,31 @@ export interface Payload {
 }
 
 
-export function abi_roundtrip(v: AbiEnvelope): AbiEnvelope;
+export function abi_roundtrip(v: AbiEnvelope<number>): AbiEnvelope<number>;
 
-export function arg_array_big(v: Envelope): void;
+export function arg_array_big(v: Envelope<number[]>): void;
 
-export function arg_array_small(v: Envelope): void;
+export function arg_array_small(v: Envelope<[number, number]>): void;
 
-export function arg_map(v: Envelope): void;
+export function arg_map(v: Envelope<Record<string, number>>): void;
 
-export function arg_option(v: Envelope): void;
+export function arg_option(v: Envelope<(number | null)>): void;
 
-export function arg_result(v: Envelope): void;
+export function arg_result(v: Envelope<({ Ok: number } | { Err: string })>): void;
 
-export function arg_tuple(v: Envelope): void;
+export function arg_tuple(v: Envelope<[number, string]>): void;
 
-export function arg_u64(v: Envelope): void;
+export function arg_u64(v: Envelope<number>): void;
 
-export function arg_vec(v: Envelope): void;
+export function arg_vec(v: Envelope<number[]>): void;
 
-export function generic_arg(v: Envelope): void;
+export function generic_arg(v: Envelope<Payload>): void;
 
-export function generic_builtin(v: Envelope): void;
+export function generic_builtin(v: Envelope<number>): void;
 
-export function generic_nested(v: Envelope): void;
+export function generic_nested(v: Envelope<Envelope<Payload>>): void;
 
-export function generic_return(v: Envelope): Envelope;
+export function generic_return(v: Envelope<Payload>): Envelope<Payload>;
 
 export function plain_arg(v: Payload): void;