+++ tests/expand/generic_constrained_struct.expanded.rs
@@ -267,22 +267,45 @@ const _: () = {
         }
     }
     #[automatically_derived]
-    impl<T: Constraint> WasmDescribe for GenericStruct<T> {
+    impl<T: Constraint> tsify::__macro_support::DescribeTsName for GenericStruct<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
-        fn describe() {
-            <Self as Tsify>::JsType::describe()
+        fn describe_ts_name() {
+            <Self as tsify::__macro_support::TsName<{ 0u8 }>>::describe_named_externref()
+        }
+        #[inline]
+        fn describe_ts_name_vector() {
+            <Self as tsify::__macro_support::TsName<
+                { 0u8 },
+            >>::describe_named_externref_vector()
         }
     }
     #[automatically_derived]
-    impl<T: Constraint> WasmDescribeVector for GenericStruct<T> {
+    impl<T: Constraint> WasmDescribe for GenericStruct<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
+        #[inline]
+        fn describe() {
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name()
+        }
+    }
+    #[automatically_derived]
+    impl<T: Constraint> WasmDescribeVector for GenericStruct<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
         fn describe_vector() {
-            <Self as Tsify>::JsType::describe_vector()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name_vector()
         }
     }
     #[automatically_derived]
     impl<T: Constraint> IntoWasmAbi for &GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -310,6 +333,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> IntoWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -321,6 +345,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> OptionIntoWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         #[inline]
@@ -331,6 +356,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> From<GenericStruct<T>> for JsValue
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         #[inline]
@@ -357,6 +383,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> VectorIntoWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         type Abi = <JsType as VectorIntoWasmAbi>::Abi;
@@ -388,6 +415,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> FromWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as FromWasmAbi>::Abi;
@@ -403,6 +431,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> OptionFromWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         #[inline]
@@ -421,6 +450,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> RefFromWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as RefFromWasmAbi>::Abi;
@@ -436,6 +466,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> VectorFromWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as VectorFromWasmAbi>::Abi;
@@ -729,22 +760,45 @@ const _: () = {
         }
     }
     #[automatically_derived]
-    impl<T: Constraint> WasmDescribe for GenericNewtype<T> {
+    impl<T: Constraint> tsify::__macro_support::DescribeTsName for GenericNewtype<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
-        fn describe() {
-            <Self as Tsify>::JsType::describe()
+        fn describe_ts_name() {
+            <Self as tsify::__macro_support::TsName<{ 0u8 }>>::describe_named_externref()
+        }
+        #[inline]
+        fn describe_ts_name_vector() {
+            <Self as tsify::__macro_support::TsName<
+                { 0u8 },
+            >>::describe_named_externref_vector()
         }
     }
     #[automatically_derived]
-    impl<T: Constraint> WasmDescribeVector for GenericNewtype<T> {
+    impl<T: Constraint> WasmDescribe for GenericNewtype<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
+        #[inline]
+        fn describe() {
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name()
+        }
+    }
+    #[automatically_derived]
+    impl<T: Constraint> WasmDescribeVector for GenericNewtype<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
         fn describe_vector() {
-            <Self as Tsify>::JsType::describe_vector()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name_vector()
         }
     }
     #[automatically_derived]
     impl<T: Constraint> IntoWasmAbi for &GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -772,6 +826,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> IntoWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -783,6 +838,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> OptionIntoWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         #[inline]
@@ -793,6 +849,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> From<GenericNewtype<T>> for JsValue
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         #[inline]
@@ -819,6 +876,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> VectorIntoWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         type Abi = <JsType as VectorIntoWasmAbi>::Abi;
@@ -850,6 +908,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> FromWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as FromWasmAbi>::Abi;
@@ -865,6 +924,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> OptionFromWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         #[inline]
@@ -883,6 +943,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> RefFromWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as RefFromWasmAbi>::Abi;
@@ -898,6 +959,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Constraint> VectorFromWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as VectorFromWasmAbi>::Abi;
@@ -1189,22 +1251,46 @@ const _: () = {
         }
     }
     #[automatically_derived]
-    impl<T: Iterator<Item = u32>> WasmDescribe for GenericAssoc<T> {
+    impl<T: Iterator<Item = u32>> tsify::__macro_support::DescribeTsName
+    for GenericAssoc<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
-        fn describe() {
-            <Self as Tsify>::JsType::describe()
+        fn describe_ts_name() {
+            <Self as tsify::__macro_support::TsName<{ 0u8 }>>::describe_named_externref()
+        }
+        #[inline]
+        fn describe_ts_name_vector() {
+            <Self as tsify::__macro_support::TsName<
+                { 0u8 },
+            >>::describe_named_externref_vector()
         }
     }
     #[automatically_derived]
-    impl<T: Iterator<Item = u32>> WasmDescribeVector for GenericAssoc<T> {
+    impl<T: Iterator<Item = u32>> WasmDescribe for GenericAssoc<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
+        #[inline]
+        fn describe() {
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name()
+        }
+    }
+    #[automatically_derived]
+    impl<T: Iterator<Item = u32>> WasmDescribeVector for GenericAssoc<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
         fn describe_vector() {
-            <Self as Tsify>::JsType::describe_vector()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name_vector()
         }
     }
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> IntoWasmAbi for &GenericAssoc<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericAssoc<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -1232,6 +1318,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> IntoWasmAbi for GenericAssoc<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericAssoc<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -1243,6 +1330,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> OptionIntoWasmAbi for GenericAssoc<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericAssoc<T>: _serde::Serialize,
     {
         #[inline]
@@ -1253,6 +1341,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> From<GenericAssoc<T>> for JsValue
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericAssoc<T>: _serde::Serialize,
     {
         #[inline]
@@ -1279,6 +1368,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> VectorIntoWasmAbi for GenericAssoc<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericAssoc<T>: _serde::Serialize,
     {
         type Abi = <JsType as VectorIntoWasmAbi>::Abi;
@@ -1310,6 +1400,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> FromWasmAbi for GenericAssoc<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as FromWasmAbi>::Abi;
@@ -1325,6 +1416,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> OptionFromWasmAbi for GenericAssoc<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         #[inline]
@@ -1343,6 +1435,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> RefFromWasmAbi for GenericAssoc<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as RefFromWasmAbi>::Abi;
@@ -1358,6 +1451,7 @@ const _: () = {
     #[automatically_derived]
     impl<T: Iterator<Item = u32>> VectorFromWasmAbi for GenericAssoc<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as VectorFromWasmAbi>::Abi;
@@ -1650,17 +1744,30 @@ const _: () = {
         }
     }
     #[automatically_derived]
+    impl<'a: 'b, 'b> tsify::__macro_support::DescribeTsName for GenericLifetime<'a, 'b> {
+        #[inline]
+        fn describe_ts_name() {
+            <Self as tsify::__macro_support::TsName<{ 0u8 }>>::describe_named_externref()
+        }
+        #[inline]
+        fn describe_ts_name_vector() {
+            <Self as tsify::__macro_support::TsName<
+                { 0u8 },
+            >>::describe_named_externref_vector()
+        }
+    }
+    #[automatically_derived]
     impl<'a: 'b, 'b> WasmDescribe for GenericLifetime<'a, 'b> {
         #[inline]
         fn describe() {
-            <Self as Tsify>::JsType::describe()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name()
         }
     }
     #[automatically_derived]
     impl<'a: 'b, 'b> WasmDescribeVector for GenericLifetime<'a, 'b> {
         #[inline]
         fn describe_vector() {
-            <Self as Tsify>::JsType::describe_vector()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name_vector()
         }
     }
     #[automatically_derived]
@@ -2103,17 +2210,30 @@ const _: () = {
         }
     }
     #[automatically_derived]
+    impl<const N: usize> tsify::__macro_support::DescribeTsName for GenericConst<N> {
+        #[inline]
+        fn describe_ts_name() {
+            <Self as tsify::__macro_support::TsName<{ 0u8 }>>::describe_named_externref()
+        }
+        #[inline]
+        fn describe_ts_name_vector() {
+            <Self as tsify::__macro_support::TsName<
+                { 0u8 },
+            >>::describe_named_externref_vector()
+        }
+    }
+    #[automatically_derived]
     impl<const N: usize> WasmDescribe for GenericConst<N> {
         #[inline]
         fn describe() {
-            <Self as Tsify>::JsType::describe()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name()
         }
     }
     #[automatically_derived]
     impl<const N: usize> WasmDescribeVector for GenericConst<N> {
         #[inline]
         fn describe_vector() {
-            <Self as Tsify>::JsType::describe_vector()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name_vector()
         }
     }
     #[automatically_derived]