+++ tests/expand/generic_struct.expanded.rs
@@ -264,22 +264,45 @@ const _: () = {
         }
     }
     #[automatically_derived]
-    impl<T> WasmDescribe for GenericStruct<T> {
+    impl<T> tsify::__macro_support::DescribeTsName for GenericStruct<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
-        fn describe() {
-            <Self as Tsify>::JsType::describe()
+        fn describe_ts_name() {
+            <Self as tsify::__macro_support::TsName<{ 0u8 }>>::describe_named_externref()
+        }
+        #[inline]
+        fn describe_ts_name_vector() {
+            <Self as tsify::__macro_support::TsName<
+                { 0u8 },
+            >>::describe_named_externref_vector()
         }
     }
     #[automatically_derived]
-    impl<T> WasmDescribeVector for GenericStruct<T> {
+    impl<T> WasmDescribe for GenericStruct<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
+        #[inline]
+        fn describe() {
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name()
+        }
+    }
+    #[automatically_derived]
+    impl<T> WasmDescribeVector for GenericStruct<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
         fn describe_vector() {
-            <Self as Tsify>::JsType::describe_vector()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name_vector()
         }
     }
     #[automatically_derived]
     impl<T> IntoWasmAbi for &GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -307,6 +330,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> IntoWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -318,6 +342,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> OptionIntoWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         #[inline]
@@ -328,6 +353,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> From<GenericStruct<T>> for JsValue
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         #[inline]
@@ -354,6 +380,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> VectorIntoWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericStruct<T>: _serde::Serialize,
     {
         type Abi = <JsType as VectorIntoWasmAbi>::Abi;
@@ -385,6 +412,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> FromWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as FromWasmAbi>::Abi;
@@ -400,6 +428,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> OptionFromWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         #[inline]
@@ -418,6 +447,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> RefFromWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as RefFromWasmAbi>::Abi;
@@ -433,6 +463,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> VectorFromWasmAbi for GenericStruct<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as VectorFromWasmAbi>::Abi;
@@ -724,22 +755,45 @@ const _: () = {
         }
     }
     #[automatically_derived]
-    impl<T> WasmDescribe for GenericNewtype<T> {
+    impl<T> tsify::__macro_support::DescribeTsName for GenericNewtype<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
-        fn describe() {
-            <Self as Tsify>::JsType::describe()
+        fn describe_ts_name() {
+            <Self as tsify::__macro_support::TsName<{ 0u8 }>>::describe_named_externref()
+        }
+        #[inline]
+        fn describe_ts_name_vector() {
+            <Self as tsify::__macro_support::TsName<
+                { 0u8 },
+            >>::describe_named_externref_vector()
         }
     }
     #[automatically_derived]
-    impl<T> WasmDescribeVector for GenericNewtype<T> {
+    impl<T> WasmDescribe for GenericNewtype<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
+        #[inline]
+        fn describe() {
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name()
+        }
+    }
+    #[automatically_derived]
+    impl<T> WasmDescribeVector for GenericNewtype<T>
+    where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
+    {
         #[inline]
         fn describe_vector() {
-            <Self as Tsify>::JsType::describe_vector()
+            <Self as tsify::__macro_support::DescribeTsName>::describe_ts_name_vector()
         }
     }
     #[automatically_derived]
     impl<T> IntoWasmAbi for &GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -767,6 +821,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> IntoWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         type Abi = <JsType as IntoWasmAbi>::Abi;
@@ -778,6 +833,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> OptionIntoWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         #[inline]
@@ -788,6 +844,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> From<GenericNewtype<T>> for JsValue
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         #[inline]
@@ -814,6 +871,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> VectorIntoWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         GenericNewtype<T>: _serde::Serialize,
     {
         type Abi = <JsType as VectorIntoWasmAbi>::Abi;
@@ -845,6 +903,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> FromWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as FromWasmAbi>::Abi;
@@ -860,6 +919,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> OptionFromWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         #[inline]
@@ -878,6 +938,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> RefFromWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as RefFromWasmAbi>::Abi;
@@ -893,6 +954,7 @@ const _: () = {
     #[automatically_derived]
     impl<T> VectorFromWasmAbi for GenericNewtype<T>
     where
+        T: tsify::__macro_support::TsName<{ 0u8 }>,
         Self: _serde::de::DeserializeOwned,
     {
         type Abi = <JsType as VectorFromWasmAbi>::Abi;